==> board_games/1_Loading_Data_csv.ipynb <==
# Data load I.

### What the code does:
* This code loads the data from TidyTuesday and creates a csv file

**Import libraries**

import pandas as pd
import requests
import json
import time
import math

## **I. Data from tidyTuesday**

####  1. Import and check data

# Import and check data
board_games_raw = pd.read_csv("https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-03-12/board_games.csv")

# Check data
pd.set_option("display.max_columns", None)
board_games_raw.head()

len(board_games_raw)

#### 2. Remove duplicates

# Number of duplicate board games
sum(board_games_raw.duplicated(subset=['name', 'year_published']))

# Remove duplicates for name + year_published
board_games_raw = board_games_raw[~board_games_raw.duplicated(subset=['name', 'year_published'])]

len(board_games_raw)

#### 3. Keep only the needed fields

# Check number of nulls
board_games_raw.isnull().sum()

# Feature list
board_games_raw.dtypes

# Remove fields with high null rate
board_games_raw = board_games_raw.loc[:, ~board_games_raw.columns.isin(['image', 'thumbnail', 'compilation', 'expansion'])]

#### 4. Data cleaning

# Lowercase name
board_games_raw['name_lower']=board_games_raw['name'].str.lower()

#### 5. Save dataset

pd.set_option("display.max_columns", None)
board_games_raw.head()

# Save dataset 
board_games_raw.to_csv("1loading_output.csv", sep=',' , quotechar='"')
