==> board_games/2_Loading_data_API.ipynb <==
# Data load II.

### What the code does:
* This code loads the data from the Board Game Atlas API and creates a csv file

**Import libraries**

import pandas as pd
import requests
import json
import time
import math

### API call options:
* In general search we can specify over 40 parameters such as: designer, publisher, maximum price etc.
* One call can return maximum 100 responses.
* If we search for name, it will return board games whose name contain the string specified (even if exact parameter is set to True).

### Technical notes:
* The full list of codes in this workbook could run for about 30-50 minutes (depending on connection to API server), most of which is from the two API calls.


# Get category list - Sample
total_url = 'https://api.boardgameatlas.com/api/search?pretty=true&client_id=6MXavgYBke'
total_response = requests.get(total_url)
total_dict = json.loads(total_response.text)


# Total number of board games listed
total_dict['count']

### Main issues:
* With the highest limit 100, it would take 1421 calls to return all games listed. 
* However, the majority of games are depreciated without too much valuable information.
* If we were to search by name, we could only get 1 game at a time.
* Therefore, we decided to look for parameters that partition the results and likely return valuable information.
* For starters, we looked into games with categories listed

**Get games by categories**

# Get category list - Sample
categ_url = 'https://api.boardgameatlas.com/api/game/categories?client_id=6MXavgYBke'
categ_response = requests.get(categ_url)
categ_dict = json.loads(categ_response.text)
list(categ_dict.values())[0][0]

# Get category ids
categ_ids = []
for cat_num in list(range(0,len(categ_dict['categories']))):
    categ_ids.append(categ_dict['categories'][cat_num]['id'])
categ_ids[1:15]

# Number of categories
len(categ_ids)

#### 2. Request board game data using the category ids 


# Sample request
categ_url = 'https://api.boardgameatlas.com/api/search?client_id=6MXavgYBke&limit=100&categories=2bdFPJUvFo'
response = requests.get(categ_url)
games_dict = json.loads(response.text)
list(games_dict.values())[0][0]

### Further issues:
* A single category may contain thousands of results.
* In theory we can skip any number results, but in practice the maximum skip value allowed is 1000.
* Therefore, with any call returning maximum 100 results, in total we can obtain up-to 1100 results from any category.


# Get board games
categ_df = pd.DataFrame()
for category_id in categ_ids:
    next_categ_url = 'https://api.boardgameatlas.com/api/search?client_id=6MXavgYBke&limit=100&categories={}'.format(category_id)
    next_response = requests.get(next_categ_url)
    next_games_dict = json.loads(next_response.text)
    categ_df = categ_df.append(pd.DataFrame(next_games_dict['games']))
    time.sleep(1)
    # This is how we specify getting 1100 results.
    if next_games_dict['count'] > 1000:
        for hundred in list(range(1,11)):
            limit = 100 * hundred
            next_categ_url = 'https://api.boardgameatlas.com/api/search?client_id=6MXavgYBke&limit=100&categories={}&skip={}'.format(
                category_id , limit)
            next_response = requests.get(next_categ_url)
            next_games_dict = json.loads(next_response.text)
            categ_df = categ_df.append(pd.DataFrame(next_games_dict['games']))
            time.sleep(1)
    elif next_games_dict['count'] > 100:
        for hundred in list(range(1,(math.ceil(next_games_dict['count'] / 100)))):
            limit = 100 * hundred
            next_categ_url = 'https://api.boardgameatlas.com/api/search?client_id=6MXavgYBke&limit=100&categories={}&skip={}'.format(
                category_id , limit)
            next_response = requests.get(next_categ_url)
            next_games_dict = json.loads(next_response.text)
            categ_df = categ_df.append(pd.DataFrame(next_games_dict['games']))
            time.sleep(1)

pd.options.display.max_columns = None
display(categ_df.head())

# Number of games 
len(categ_df)

# Feature list
categ_df.dtypes

#### 3. Keep only the needed fields

# Keep only useful fields
categ_df_short = categ_df.loc[:, categ_df.columns.isin([
    'artists',
    'average_learning_complexity',
    'average_strategy_complexity',
    'average_user_rating',
    'categories',
    'comment_count',
    'description',
    'description_preview',
    'designers',
    'discount',
    'id',
    'is_historical_low',
    'listing_clicks',
    'max_players',
    'max_playtime',
    'min_age',
    'min_players',
    'min_playtime',
    'name',
    'num_user_ratings',
    'players',
    'plays',
    'playtime',
    'price',
    'price_au',
    'price_ca',
    'price_uk',
    'primary_publisher',
    'publishers',
    'visits',
    'year_published'  
])]

categ_df_short.isnull().sum()

#### 4. Data cleaning

# Lowercase name
categ_df_short['name_lower']=categ_df_short['name'].str.lower()

# Remove original fields
categ_df_short = categ_df_short.loc[:, ~categ_df_short.columns.isin(['artists', 'primary_publisher', 'categories', 'designers', 'publishers', 'description', 'description_preview'])]

# Feature list
categ_df_short.dtypes

pd.options.display.max_columns = None
categ_df_short.head()

#### 5. Remove duplicates

# Remove duplicates
no_duplicate_categ_df = categ_df_short[~categ_df_short.duplicated(subset=['name_lower', 'year_published'])]

len(no_duplicate_categ_df)

pd.options.display.max_columns = None
no_duplicate_categ_df.head()

#### 5. Load More data

### Considerations:
* So far we have less names than in the original csv and we haven't matched names yet.
* To increase data we tried the second categorical variables in the API call, which is the list of mechanics in a game.

# The call is very similar, but this time we only select the fields we know we will use, you can specify this at the API call.

field_list = ""
for list_item in list(categ_df_short.columns):
    field_list = field_list + "," + list_item
field_list = field_list[1:]
field_list = field_list[:-11]
field_list

# Get mechanics list 
mechanics_url = 'https://api.boardgameatlas.com/api/game/mechanics?pretty=true&client_id=6MXavgYBke'
mechanics_response = requests.get(mechanics_url)
mechanics_dict = json.loads(mechanics_response.text)

# Get mechanics ids
mechanics_ids = []
for mech_num in list(range(0,len(mechanics_dict['mechanics']))):
    mechanics_ids.append(mechanics_dict['mechanics'][mech_num]['id'])

# Get board games
mech_df = pd.DataFrame()
for mech_id in mechanics_ids:
    next_mech_url = 'https://api.boardgameatlas.com/api/search?client_id=6MXavgYBke&limit=100&mechanics={}&fields={}'.format(mech_id,field_list)
    next_response = requests.get(next_mech_url)
    next_games_dict = json.loads(next_response.text)
    mech_df = mech_df.append(pd.DataFrame(next_games_dict['games']))
    time.sleep(1)
    if next_games_dict['count'] > 1000:
        for hundred in list(range(1,11)):
            limit = 100 * hundred
            next_mech_url = 'https://api.boardgameatlas.com/api/search?client_id=6MXavgYBke&limit=100&mechanics={}&skip={}&fields={}'.format(mech_id , limit, field_list)
            next_response = requests.get(next_mech_url)
            next_games_dict = json.loads(next_response.text)
            mech_df = mech_df.append(pd.DataFrame(next_games_dict['games']))
            time.sleep(1)
    elif next_games_dict['count'] > 100:
        for hundred in list(range(1,(math.ceil(next_games_dict['count'] / 100)))):
            limit = 100 * hundred
            next_mech_url = 'https://api.boardgameatlas.com/api/search?client_id=6MXavgYBke&limit=100&mechanics={}&skip={}&fields={}'.format(mech_id , limit, field_list)
            next_response = requests.get(next_mech_url)
            next_games_dict = json.loads(next_response.text)
            mech_df = mech_df.append(pd.DataFrame(next_games_dict['games']))
            time.sleep(1)

# Add lower name
mech_df_names = mech_df.copy()
mech_df_names['name_lower'] = mech_df_names['name'].str.lower()


# Remove duplicates

print(len(mech_df_names))
no_duplicate_mech_df = mech_df_names[~mech_df_names.duplicated(subset=['name_lower', 'year_published'])]
no_duplicate_mech_df = no_duplicate_mech_df.reset_index()
len(no_duplicate_mech_df)

#### 6. Finalizing results

### Other options:
* We managed to get a lot more results with the second query
* Our final option would be to request make individual requests based on game names, but it would take too long.
* Therefore, the last step is to join the category and mechnism tables and save the results.


board_games_api = pd.concat( [no_duplicate_categ_df, no_duplicate_mech_df])
pd.options.display.max_columns = None
board_games_api.head()


# Remove duplicates

print(len(board_games_api))
board_games_api = board_games_api[~board_games_api.duplicated(subset=['id'])]
len(board_games_api)

# Save dataset 
board_games_api.to_csv("2loading_output.csv", sep=',' , quotechar='"')

### Final note:
* Our approach to API calls netted over 20k results under 30 minutes which is much faster than individual requests.
* However, our selection was not random, we could only query the first 1100 results of any category/mechanism.
* Upon further investigation, we found that results are ordered by relevance and the ending pages we could not access contain
mainly missing information.
* Consider the last page of [card games category](https://www.boardgameatlas.com/category/eX8uuNlQkQ/card-game?skip=1900&categories=eX8uuNlQkQ): most prices and other relevant information are missing.
* As you will see in further notebooks, we would likely discard most of these results for our analysis, therefore the issue of representation would not be resolved with more APi calls.
