==> board_games/3_Merging_data.ipynb <==
## **III. Merge csv and API**

### What the code does:
* This code merges the datasets from TidyTuesday and the Board Game Atlas API and creates the workfile

**Import libraries**

import pandas as pd
import requests
import json
import time
import math

### Technical notes:
* The full list of codes in this workbook run for about 15 minutes, most of which is the fuzzy merge.
* For the fuzzy merge rapidfuzz module is required.


**Load in datasets**

# Import dataset - Tidytuesday
board_games = pd.read_csv("1loading_output.csv", sep=',')
board_games.columns = [str(col) + '_base' for col in board_games.columns]
board_games.head()

board_games_api = pd.read_csv("2loading_output.csv", sep=',')
board_games_api.columns = [str(col) + '_api' for col in board_games_api.columns]
board_games_api.head()

**Try simple join based on names**

# Merge datasets on name and year_published
united_df = pd.merge(board_games, 
                     board_games_api, 
                     how='inner', 
                     left_on=['name_lower_base', 'year_published_base'], 
                     right_on=['name_lower_api', 'year_published_api'])
len(united_df)

### Improvements, approximate match:
* The two API calls already made good results, almost half the csv games were found.
* But maybe we can do better by approximate ("fuzzy") search.


**Import libraries**

# !pip install rapidfuzz -q # fuzzy search package, install takes about 2 minutes
# from rapidfuzz import fuzz # main score package, not required
from rapidfuzz.process import extractOne, extract, extract_iter # get best match, get all matches, iterated
from rapidfuzz.string_metric import levenshtein, normalized_levenshtein # more complex match scores
# from rapidfuzz.fuzz import ratio #simple ratio, not required

board_games['name_lower_base'][0]

# example of finding all matches above 85% threshold
# Results are not very good
for x in extract_iter(board_games['name_lower_base'][0],board_games_api['name_lower_api'],score_cutoff = 85):
    print(x)

# Examples of best match above threshold.
# This is much better
for game in board_games['name_lower_base'][0:15]:
    best_find = extractOne(game,board_games_api['name_lower_api'],score_cutoff = 85)
    print(game)
    if not (best_find is None):
        print(best_find[0])
        print(best_find[1])
    else:
        print('No match')

# runs for about 10 minutes at 20000 rows
# get best match with score for all board games from raw file if there is any match of at least 85% score.
import numpy as np


matches = []
scores = []

for game in board_games['name_lower_base']:
    best_find = extractOne(game,board_games_api['name_lower_api'],score_cutoff = 85)
    if not (best_find is None):
        matches.append(best_find[0])
        scores.append(best_find[1])
    else:
        matches.append(np.nan)
        scores.append(np.nan)


# Create data.frame of scores

aux_df = pd.DataFrame({'csv': pd.Series(board_games['name_lower_base'] ), 'match': matches , 'score': scores})
aux_df

# Percentage of matches found
(len(aux_df) - sum(aux_df['match'].isnull()))/len(aux_df)

### Results:
* Over 93% of games were found a match, which seems a bit too optimistic.
* Also, some of the matches with scores between 85-95% are not very good.
* However, from the same module we imported we can use levenshtein scores which are more precise and can be customized.


**Examples of lower levenshtein scores, setting threshold at 80% seems good.**

normalized_levenshtein('space beans', 'space base',weights=(1,1,1))

normalized_levenshtein('taki', 'tak',weights=(1,1,1))

**However, longer names seem to be more complex.**

normalized_levenshtein('battle of the bulge','battle of the bulge 1944')

# We can set up weights.

normalized_levenshtein('battle of the bulge','battle of the bulge 1944',weights=(1,1,1)) # by default it has weights 1-1-1

# Weights stand for: 1 penalty for 1 digit additional - 1 digit less - 1 digit substituted

# Weights can be changed.

normalized_levenshtein('battle of the bulge','battle of the bulge 1944',weights=(1,1,2))

# It seem that for longer names, substitution is a bigger issuer than less/more words.

### Rules:
* Some board games have really long names, where substitutions are the biggest issue, whereas additional words are not so much.
* Examples such as: "third edition" , "collectible" are not as good indicators of false match.
* So for longer names we penalize only substitutions heavily.
* As tested, we will use 80% levenshtein match as threshold for true match.


weighted_score = []
for game in list(range(0,10530)):
    if not (aux_df['match'][game] is np.nan):
        if len(aux_df['match'][game]) > 15:
            weighted_score.append(normalized_levenshtein(aux_df['csv'][game], aux_df['match'][game],weights=(1,1,2)))
        else:
            weighted_score.append(normalized_levenshtein(aux_df['csv'][game], aux_df['match'][game],weights=(1,1,1)))
    else:
        weighted_score.append(np.nan)

aux_df['weighted_score'] = weighted_score
aux_df['threshold'] = np.where(aux_df['weighted_score']>=80,True,False)
print(sum(aux_df['threshold'])) # number of matches above threshold

### Conclusions:
* Using fuzzy search with relatively conservative scores netted us over 15% more matches.
* Finally: merge auxilary results table, keep only true matches.
* Based on true matches, merge to api table.
* Remove technical columns.


# Add results to table to merge based on it
united_df = board_games.join(aux_df[['match','threshold']])
# merge on index as same length since we found something for each game
united_df

# Filter out only matches above threshold
united_df = united_df[united_df['threshold'] == True] 

# Merge datasets on name and match found
# no_duplicate_board_games_api = board_games_api[~board_games_api.duplicated(subset=['name_lower_api', 'year_published_api'])]

united_df = pd.merge(united_df, 
                     board_games_api, 
                     how='inner', 
                     left_on=['match'], 
                     right_on=['name_lower_api'])
print(len(united_df))
# Remove duplicates from multiple matches
united_df = united_df[~united_df.duplicated(subset=['name_lower_base', 'year_published_base'])]
len(united_df)

united_df.columns

united_df['Unnamed: 0'] = united_df.index # Adding new index
# Move it first
cols = united_df.columns.tolist()
cols = cols[-1:] + cols[:-1]
united_df = united_df[cols] 
united_df = united_df.drop(['match', 'threshold','index_api'], axis = 1) # Removing technical columns


united_df.columns

# Save dataset 
united_df.to_csv("board_games_workfile.csv", sep=';' , quotechar='"')
